# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "It's A Wrap POS API",
        "description": "API for It's A Wrap POS system - supports Flutter mobile apps and web applications",
        "contact": {
            "name": "API Support",
            "email": "support@itsawrap.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "List all orders",
                "description": "Get all orders with optional filtering",
                "operationId": "8f3183bfa0ecb8a2c864cbf972b8061d",
                "parameters": [
                    {
                        "name": "status_id",
                        "in": "query",
                        "description": "Filter by status",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "session_id",
                        "in": "query",
                        "description": "Filter by cash session",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "customer_id",
                        "in": "query",
                        "description": "Filter by customer",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Filter by date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of orders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Order"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create an order",
                "description": "Create a new order with items",
                "operationId": "b6490f9d9026dbe871b41207e118fffb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status_id",
                                    "subtotal",
                                    "total"
                                ],
                                "properties": {
                                    "number": {
                                        "type": "string",
                                        "example": "ORD-001",
                                        "nullable": true
                                    },
                                    "customer_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "status_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "subtotal": {
                                        "type": "number",
                                        "example": 25.99
                                    },
                                    "discount": {
                                        "type": "number",
                                        "example": 0,
                                        "nullable": true
                                    },
                                    "discount_percent": {
                                        "type": "number",
                                        "example": 0,
                                        "nullable": true
                                    },
                                    "service_charge": {
                                        "type": "number",
                                        "example": 0
                                    },
                                    "total": {
                                        "type": "number",
                                        "example": 25.99
                                    },
                                    "comments": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "is_delivery": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "is_reward": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "item_id": {
                                                    "type": "integer"
                                                },
                                                "price": {
                                                    "type": "number"
                                                },
                                                "quantity": {
                                                    "type": "integer"
                                                },
                                                "options": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "option_value_id": {
                                                                "type": "integer"
                                                            },
                                                            "price": {
                                                                "type": "number"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get an order",
                "description": "Get a single order with all details",
                "operationId": "79c1e32fc3603170dfd7659562133aeb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update an order",
                "description": "Update an existing order",
                "operationId": "be1c75741bfcbee84e078c24a476f6ee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status_id": {
                                        "type": "integer"
                                    },
                                    "comments": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "discount": {
                                        "type": "number",
                                        "nullable": true
                                    },
                                    "discount_percent": {
                                        "type": "number",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Order not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Delete an order",
                "description": "Delete an order",
                "operationId": "79a357ed1e1e8bdc55b7df96efcddf18",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Order deleted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Order": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "customer_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "cash_session_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "status_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "order_number": {
                        "type": "string",
                        "example": "ORD-001"
                    },
                    "subtotal": {
                        "type": "number",
                        "format": "float",
                        "example": 25.99
                    },
                    "tax_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 2.14
                    },
                    "discount_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 0
                    },
                    "total": {
                        "type": "number",
                        "format": "float",
                        "example": 28.13
                    },
                    "notes": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}